statically link game, cgame and ui
about 20% into this commit I realized I should have split it into dozens of smaller commits, but didn't want to undo my work up to that point, so now I'm left with this monolithic monster... I'm sure I introduced bugs, so if you find this commit through git bisect, I'm so sorry.

note that this is not actually tested in any way yet - I merely made it compile again and then called it a day, and thought I should backup my weeks worth of changes in the cloud before doing anything else.

as for what I did: there was a bunch of shared code (bg_*) that behaved differently based on defines depending on whether it was compiled as cgame, game or ui. I was left with the choice of compiling all that code repeatedly in different namespaces or using templates and opted for the latter, as it makes everything more explicit.

as a result the code now feels a lot more ugly - I hope I can fix that in time.